Containers rewrite (#312)
* Split container data to simplify containers setup

* Fix object definitions test

* Add containers wrapper class

* Replace container predicate with item restrictions

* Add replace stack mode with stacking rule

* Move item removal into its own check class and remove secondary field from container

* Add container transaction system

* Add basic transaction operators and unit tests

* Make transactions one-to-one with containers

* Extract State and Change managers from Transaction

* Remove helper functions from transaction

* Rewrite TransactionTest

* Add more operation functions and refactor unit tests

* Return existing transaction when linking containers a second time

* Rewrite remaining operation unit tests

* Tidy up ItemStackingRule

* Disable item charges #311

* Add replacing item operation

* Add shift insert item operation

* Replace transaction valid checks with item rule and removal checks

* Remove redundant transaction error setter

* Transaction documentation

* Rename quantity to amount

* Replace some container valid checks with rule and removal checks

* Convert replace operation

* Convert swap operation

* Convert move all operation

* Convert move operation

* Fix item restriction rule changes

* Converted move operations and shop restrictions

* Remove unused insert

* Convert remove

* Convert clear

* Convert add

* Remove unused container operations

* Convert set

* Remove container result

* Remove container item definitions

* Move bank sort out of container

* Add remove operation default amount

* Make container getCount return int by default

* Remove purchase method

* Replace transaction error nullable with No error

* Shorten transaction link method name

* Add container rule functions and start simplifying container

* Remove container valid checks

* Renamed has(item, banked = true) to hasBanked

* Refactor getItem to get operator

* Refactor getId and getAmount to id and amount

* Replace getId and getAmount with get calls

* Refactor getCount to count

* Misc improvements, move Events from container into ChangeManager

* Add from and to containers to ItemChanged update

* Fix bank withdrawal

* Remove bank sorting in favour of proper removal

* Fixed can't be noted bank message

* Fix interface switch packet

* Replace shiftInsert with regular shift operation

* Refactor bank tabs

* Add optional different add id to moveToLimit

* Improve moveToLimit item order

* Fix shops and don't wipe errors on linked transactions

* Fix item defs in world integration tests

* Optimise sendContainerItems a bit

+++ game/src/main/kotlin/world/gregs/voidps/world/community/trade/lend/Loan.kt
@@ -3,9 +3,7 @@ package world.gregs.voidps.world.community.trade.lend
 import com.github.michaelbull.logging.InlineLogger
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.entity.*
-import world.gregs.voidps.engine.entity.character.contain.beastOfBurden
-import world.gregs.voidps.engine.entity.character.contain.equipment
-import world.gregs.voidps.engine.entity.character.contain.inventory
+import world.gregs.voidps.engine.entity.character.contain.*
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.definition.ItemDefinitions
 import world.gregs.voidps.engine.tick.delay
@@ -57,7 +55,7 @@ object Loan {
     fun getTimeRemaining(player: Player, timeKey: String): Long {
         return when {
             player.contains(timeKey) -> {
-                val timeout: Int = player.getOrNull(timeKey) ?: return -1
+                val timeout: Long = player.getOrNull(timeKey) ?: return -1
                 System.currentTimeMillis() - timeout
             }
             else -> 0
